Add files via upload
wrote a program in F# that when passing two text files to cat, e.g. a.txt and b.txt, then the program
prints the contents of file a.txt followed by the contents of b.txt to the screen. Consider an inverse
version of cat, called tac, which prints the files in reverse order and prints each file from the last
to the first character.

The function CountLinks should read the page given in url and count how many links that page has to other
pages.